Develop (#8)
* formatting

* creating components for InputText

* adding publicPath for the ooutput file

* adding components

* updated packages

* adding nodemon and webpack settings

* adding more configuration for node js and webpack

* adding view, edit, and list for user route

* added user create component

* added middleware for delete request

* fixing some errors for user list

* Create package-lock.json

* if condition for no users

* fixing null value for empty value

* user routes file added, and added dinamic rendering

* fixing styling for buttons

* cleaning styling for components

* fixing key for a row withing the breadcrumbs

* React Bootstrap Modal added to common folder

* new route added for signing up users

* adding api end point for signup service

* adding styling for sign up and sign in, also changing route for login

* authentication added to front end against back end

* showing message

* fix url string

* adding auth for token and avatar

* avatar component

* adding styling to the header and fixing the item for the avatar dropdown

* optima font family added

* checkin token for avatar

* converting Navbar to const functional component

* updating packages

* adding styling to nav bar

* Optima font-family added to the body tag style

* removing underline from nav-items

* adding styling to breadcrumbs

* hashing paswords when creating users

* cleaning code

* cleaning code and making comments about next steps for verify tokens

* express added, weird because it was working

* href to another page after signing in and signin out

* rearrange items in the nav bar

* handling signout in avatar dropdown

* Sidebar component created and improved header and navbar styling

* adding axios request interceptors

* generating tokens in the back end

* auth back end

* handling sign up, sign in, create user, update user

* styling the header and nav bar

* more styling

* fixing styling

* adding delete icon, arraUtils, and converted from class to functional component

* changing to js file

* fixing infinite loop in useEffect

+++ src/components/users/UserEditEntry.jsx
@@ -1,8 +1,9 @@
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 import './Users.css';
 import UserFormEntry from './UserFormEntry';
 import UserService from "./UserService";
 import Loading from "../common/Loading";
+import {confirm} from "../common/modal/confirm";
 
 export default class UserEditEntry extends React.Component {
 
@@ -10,7 +11,9 @@ export default class UserEditEntry extends React.Component {
 
     constructor(props) {
         super(props);
-        this.state = {};
+        this.state = {
+            modalShow: false
+        };
     }
 
     componentDidMount() {
@@ -35,18 +38,13 @@ export default class UserEditEntry extends React.Component {
 
     handleDelete = (e) => {
         const userId = this.state.user.userId;
-
-        this.userService.deleteUser(userId).then(data => {
-            window.location.pathname = `/users`;
-        });
-
-        // confirm("Are you sure you want to delete this user?", (res) => {
-        //     if (res) {
-        //         this.http.delete(`/api/edit/user/${this.props.itemSummaries.user.id}`).then(res => {
-        //             window.location.pathname = `/admin/user/admin/all`;
-        //         });
-        //     }
-        // })
+        confirm("Are you sure you want to delete this user?", (res) => {
+            if (res) {
+                this.userService.deleteUser(userId).then(data => {
+                    window.location.pathname = `/users`;
+                });
+            }
+        })
     }
 
     isFormValid = () => {
@@ -72,7 +70,7 @@ export default class UserEditEntry extends React.Component {
             return <Loading />
         }
         return (
-            <div style={{paddingTop: "40px"}}>
+            <div style={{paddingTop: "20px"}}>
                 <h1>Edit User</h1>
                 <UserFormEntry user={this.state.user} onChange={this.onFormChange}/>
                 <div className="form-group">
